+++ comb_spec_searcher/specification.py
@@ -71,6 +71,7 @@ class CombinatorialSpecification(
         self.root = root
         self.rules_dict: Dict[CombinatorialClassType, AbstractRule] = {}
         self._populate_rules_dict(strategies, equivalence_paths, expand_verified)
+        self._remove_redundant_rules()
         for rule in list(
             self.rules_dict.values()
         ):  # list as we lazily assign empty rules
@@ -135,6 +136,28 @@ class CombinatorialSpecification(
                         rule = equivalence_rules[(b, a)].to_reverse_rule()
                     rules.append(rule)
                 self.rules_dict[start] = EquivalencePathRule(rules)
+
+    def _remove_redundant_rules(self) -> None:
+        rules_dict = copy(self.rules_dict)
+
+        def prune(comb_class: CombinatorialClassType) -> None:
+            try:
+                rule = rules_dict.pop(comb_class)
+            except KeyError:
+                assert comb_class in self.rules_dict or comb_class.is_empty()
+                return
+            if isinstance(rule, EquivalencePathRule):
+                try:
+                    rule = rules_dict.pop(rule.children[0])
+                except KeyError:
+                    assert comb_class in self.rules_dict
+                    return
+            for child in rule.children:
+                prune(child)
+
+        prune(self.root)
+        for rule in rules_dict.values():
+            self.rules_dict.pop(rule.comb_class)
 
     def _expand_verified_comb_classes(
         self, verification_packs: Dict[CombinatorialClassType, StrategyPack],
@@ -399,6 +422,7 @@ class CombinatorialSpecification(
             return res
 
         res = update_res(self.root, res)
+        assert not rules_dict
         return res + "\n"
 
     def equations_string(self) -> str: